django_task/pyproject.toml
[tool.poetry]
name = "django-task"
version = "0.1.0"
description = ""
authors = ["Unti1 <kostyuchenko.work@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
django-debug-toolbar = "^4.3.0"
psycopg2 = "^2.9.9"
django-mptt = "^0.16.0"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
